+++ pointer_queue_test.go
@@ -4,14 +4,14 @@ import "testing"
 
 func BenchmarkPointerQueue_Enqueue(b *testing.B) {
 	queue := NewPointerQueue()
-	for i := uint32(0); i < uint32(b.N); i++ {
+	for i := uint64(0); i < uint64(b.N); i++ {
 		queue.Enqueue(i)
 	}
 }
 
 func BenchmarkPointerQueue_Dequeue(b *testing.B) {
 	queue := NewPointerQueue()
-	for i := uint32(0); i < uint32(b.N); i++ {
+	for i := uint64(0); i < uint64(b.N); i++ {
 		queue.Enqueue(i)
 	}
 	b.ResetTimer()
@@ -22,17 +22,17 @@ func BenchmarkPointerQueue_Dequeue(b *testing.B) {
 
 func TestPointerQueue_Enqueue(t *testing.T) {
 	queue := NewPointerQueue()
-	for i := uint32(0); i < 3000; i++ {
+	for i := uint64(0); i < 3000; i++ {
 		queue.Enqueue(i)
 	}
 }
 
 func TestPointerQueue_Dequeue(t *testing.T) {
 	queue := NewPointerQueue()
-	for i := uint32(0); i < 3000; i++ {
+	for i := uint64(0); i < 3000; i++ {
 		queue.Enqueue(i)
 	}
-	for i := uint32(0); i < 3000; i++ {
+	for i := uint64(0); i < 3000; i++ {
 		k, ok := queue.Dequeue()
 		if k != i || !ok {
 			t.Fatalf("Item %d couldn't be dequeued ok=%t v=%d", i, ok, k)
